Add a new field `message` to config.json
The greeting message can now be set to a custom static message.

The field in the config required is `message` and it needs to be a
string. If the string is not empty, it will be shown instead of the
default greet message.

In order to disbale it, the field needs to be omitted or set to an empty
string like `""`.